backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-appconf==1.0.5
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fire==0.5.0
idna==3.4
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
urllib3==2.0.4
whitenoise==6.5.0
